Refactor:all_resource
