+++ app/Http/Controllers/Api/Dashboard/RoomController.php
@@ -9,7 +9,7 @@ use Illuminate\Http\Request;
 use App\Http\Helpers\ApiResponse;
 use Illuminate\Support\Facades\Log;
 use App\Http\Controllers\Controller;
-use App\Http\Resources\Api\RoomResource;
+use App\Http\Resources\Api\Dashboard\RoomResource;
 use Exception;
 use Illuminate\Support\Facades\Validator;
